ekf2: move GPS control logic to separate source file
