+++ src/modules/ekf2/CMakeLists.txt
@@ -54,6 +54,7 @@ px4_add_module(
 		EKF/EKFGSF_yaw.cpp
 		EKF/estimator_interface.cpp
 		EKF/gps_checks.cpp
+		EKF/gps_control.cpp
 		EKF/gps_fusion.cpp
 		EKF/gps_yaw_fusion.cpp
 		EKF/imu_down_sampler.cpp
