+++ src/modules/ekf2/EKF/CMakeLists.txt
@@ -42,6 +42,7 @@ add_library(ecl_EKF
 	EKFGSF_yaw.cpp
 	estimator_interface.cpp
 	gps_checks.cpp
+	gps_control.cpp
 	gps_fusion.cpp
 	gps_yaw_fusion.cpp
 	imu_down_sampler.cpp
